Re-centralize models code in models module
The models import was broken for the HTTP application.

Also keep the two scripts as they are cleaner but use non-relative imports,
which might fails in case of module name conflict.
