+++ .isort.cfg
@@ -4,4 +4,4 @@ include_trailing_comma=True
 force_grid_wrap=0
 use_parentheses=True
 line_length=88
-known_third_party = dateutil,flask,hglib,imblearn,jsone,libmozdata,numpy,pandas,pytest,requests,responses,rs_parsepatch,setuptools,shap,sklearn,taskcluster,tqdm,xgboost,yaml,zstandard
+known_third_party = dateutil,flask,hglib,imblearn,jsone,libmozdata,models,numpy,pandas,pytest,requests,responses,rs_parsepatch,setuptools,shap,sklearn,taskcluster,tqdm,xgboost,yaml,zstandard
